re-structure of code, to smaller parts in order to make it easier to read
