+++ .gitignore
@@ -3,3 +3,6 @@
 .DS_store
 /__pycache__
 /fagskole-env/
+views/__pycache__/contacts_view.cpython-313.pyc
+views/__pycache__/inventory_view.cpython-313.pyc
+views/__pycache__/orders_view.cpython-313.pyc
